I plugged the [range-sse algorithm](https://github.com/cyb70289/utf8/blob/master/range-sse.c) so we can compare the performances.

The author claims it should not bring any significant improvement on sse. Seems to be true.

Before:
number of iterations 1000 
number of bytes 631515 number of structural chars 55264 ratio 0.088
mem alloc instructions:       1256 cycles:        911 (0.07 %) ins/cycles: 1.38 mis. branches:          5 (cycles/mis.branch 153.85) cache accesses:      30452 (failure          0)
 mem alloc runs at 0.00 cycles per input byte.
stage 1 instructions:    3192517 cycles:     857482 (62.58 %) ins/cycles: 3.72 mis. branches:       1608 (cycles/mis.branch 533.04) cache accesses:      30452 (failure          2)
 stage 1 runs at 1.36 cycles per input byte.
stage 2 instructions:    1577387 cycles:     511731 (37.35 %) ins/cycles: 3.08 mis. branches:       1662  (cycles/mis.branch 307.82)  cache accesses:      55074 (failure          8)
 stage 2 runs at 0.81 cycles per input byte and 9.26 cycles per structural character.
 all stages: 2.17 cycles per input byte.
Estimated average frequency: 4.646 GHz.
Min:  0.000294926 bytes read: 631515 Gigabytes/second: 2.14127

With range-sse:
number of iterations 1000 
number of bytes 631515 number of structural chars 55264 ratio 0.088
mem alloc instructions:       1256 cycles:        885 (0.07 %) ins/cycles: 1.42 mis. branches:          5 (cycles/mis.branch 171.05) cache accesses:      29328 (failure          1)
 mem alloc runs at 0.00 cycles per input byte.
stage 1 instructions:    3094438 cycles:     825148 (61.60 %) ins/cycles: 3.75 mis. branches:       1496 (cycles/mis.branch 551.29) cache accesses:      29328 (failure         62)
 stage 1 runs at 1.31 cycles per input byte.
stage 2 instructions:    1577387 cycles:     513460 (38.33 %) ins/cycles: 3.07 mis. branches:       1767  (cycles/mis.branch 290.56)  cache accesses:      52921 (failure        185)
 stage 2 runs at 0.81 cycles per input byte and 9.29 cycles per structural character.
 all stages: 2.12 cycles per input byte.
Estimated average frequency: 4.636 GHz.
Min:  0.000288942 bytes read: 631515 Gigabytes/second: 2.18561

A neon version might come soon.