@lemire 2019-08-14
At a glance, this looks like progress...
You go from 3192517 instructions down to 3094438 instructions in stage 1... that's a 3% improvement!

Note that this is 3% better over all of stage 1... whereas the new code only affects some of it. That's great stuff if it holds up... gaining 3% would be fantastic.

Which file did you use? twitter.json?