@lemire 2019-11-21
Made obsolete by https://github.com/lemire/simdjson/pull/365